Minor internal changes

+++ lightcurvedb/storage/postgres/schema.py
@@ -23,12 +23,12 @@ CREATE TABLE flux_measurements (
     band_name TEXT NOT NULL REFERENCES bands(name),
     source_id INTEGER NOT NULL REFERENCES sources(id),
     time TIMESTAMPTZ NOT NULL,
-    ra DOUBLE PRECISION NOT NULL CHECK (ra >= -180 AND ra <= 180),
-    dec DOUBLE PRECISION NOT NULL CHECK (dec >= -90 AND dec <= 90),
-    ra_uncertainty DOUBLE PRECISION,
-    dec_uncertainty DOUBLE PRECISION,
-    i_flux DOUBLE PRECISION NOT NULL,
-    i_uncertainty DOUBLE PRECISION,
+    ra REAL NOT NULL CHECK (ra >= -180 AND ra <= 180),
+    dec REAL NOT NULL CHECK (dec >= -90 AND dec <= 90),
+    ra_uncertainty REAL,
+    dec_uncertainty REAL,
+    i_flux REAL NOT NULL,
+    i_uncertainty REAL,
     extra JSONB,
     PRIMARY KEY (source_id, id)
 ) PARTITION BY HASH (source_id);
